cleaned up the code

+++ src/m3dvc.cpp
@@ -29,14 +29,11 @@ m3dvc::~m3dvc() {
 void m3dvc::mergeVisibleDepthData() {
   int indexForDepthBlock = 0;
   int length = (width*height) / 2;
-  for (int i = 0; i < length; i++)
-  {
-    if (i >= length / 2)
-    {
+  for (int i = 0; i < length; i++) {
+    if (i >= length / 2) {
       encoded_merge_visible_depth_data[i] = encoded_depth_data[indexForDepthBlock];
       indexForDepthBlock++;
-    } else
-    {
+    } else {
       encoded_merge_visible_depth_data[i] = encoded_reduced_visible_data[i];
     }
   }
@@ -100,8 +97,7 @@ int* m3dvc::getLowFreqArrayExt(double* dwtOrgImage, int Width, int Height) {
   //creating low freq
   int j = 0;
   int index = 0;
-  for (int i = 0; i < Width*Height; i = i + 2)
-  {
+  for (int i = 0; i < Width*Height; i = i + 2) {
     oneMergeFilter[index] = ((int)dwtOrgImage[i] + (int)dwtOrgImage[i + 1]) / 2;
     oneMergeFilter[index + (Width*Height / 2)] = (int)dwtOrgImage[i] - oneMergeFilter[index];
     j++;
@@ -122,13 +118,10 @@ int* m3dvc::getEncodedDWTExt(double* dwtOrgImage, int Width, int Height) {
 
   //merging high and low freq into one
   int j = 0;
-  for (int i = 0; i < Width*Height; i++)
-  {
-    if (i < (Width*Height) / 2)
-    {
+  for (int i = 0; i < Width*Height; i++) {
+    if (i < (Width*Height) / 2) {
       oneMergeFilter[i] = lowPassFilter[i];
-    } else
-    {
+    } else {
       oneMergeFilter[i] = highPassFilter[j];
       j++;
     }
@@ -149,20 +142,16 @@ void m3dvc::M3DVC_visible_encode(int* dwtOrgImage) {
 
   //merging high and low freq into one
   int j = 0;
-  for (int i = 0; i < width*height; i++)
-  {
-    if (i < (width*height) / 2)
-    {
+  for (int i = 0; i < width*height; i++) {
+    if (i < (width*height) / 2) {
       encoded_visible_data[i] = lowPassFilter[i];
-    } else
-    {
+    } else {
       encoded_visible_data[i] = highPassFilter[j];
       j++;
     }
   }
 
-  for (int i = 0; i < (int)(width*height / 4); i++)
-  {
+  for (int i = 0; i < (int)(width*height / 4); i++) {
     int indexY = i / (width / 2);
     int indexX = i % (width / 2);
 
@@ -186,20 +175,16 @@ void m3dvc::M3DVC_visible_encode(double* dwtOrgImage) {
 
   //merging high and low freq into one
   int j = 0;
-  for (int i = 0; i < width*height; i++)
-  {
-    if (i < (width*height) / 2)
-    {
+  for (int i = 0; i < width*height; i++) {
+    if (i < (width*height) / 2) {
       encoded_visible_data[i] = lowPassFilter[i];
-    } else
-    {
+    } else {
       encoded_visible_data[i] = highPassFilter[j];
       j++;
     }
   }
 
-  for (int i = 0; i < (int)(width*height / 4); i++)
-  {
+  for (int i = 0; i < (int)(width*height / 4); i++) {
     int indexY = i / (width / 2);
     int indexX = i % (width / 2);
 
@@ -281,15 +266,12 @@ void m3dvc::M3DVC_visible_encode(double* dwtOrgImage) {
 void m3dvc::extractVisibleDepthData(char* receivedData) {
   int indexT = 0;
   int indexT2 = 0;
-  for (int i = 0; i < (width*height) / 2; i++)
-  {
+  for (int i = 0; i < (width*height) / 2; i++) {
     uint8_t temp = (uint8_t)receivedData[i];
-    if (i < (width*height) / 4)
-    {
+    if (i < (width*height) / 4) {
       decoded_reduced_visible_data[indexT] = temp;
       indexT++;
-    } else
-    {
+    } else {
       encoded_depth_data[indexT2] = temp;
       indexT2++;
     }
@@ -299,8 +281,7 @@ void m3dvc::extractVisibleDepthData(char* receivedData) {
 void m3dvc::getReplicateBlocks(int compressedIndicator) {
   //	int* oneMergeFilter=new int[Width*Height];
 
-  for (int i = 0; i < (width*height / compressedIndicator); i++)
-  {
+  for (int i = 0; i < (width*height / compressedIndicator); i++) {
     int indexY = i / (width / 2);
     int indexX = i % (width / 2);
 
@@ -318,12 +299,11 @@ void m3dvc::getReplicateBlocks(int compressedIndicator) {
 void m3dvc::M3DVC_visible_decode() {
   //	int* dwtRecoveredImage=new int[Width*Height];
 
-  for (int i = 0; i < (width*height) / 2; i++)
-  {
+  for (int i = 0; i < (width*height) / 2; i++) {
     double recoverVal1 = decoded_visible_data[i] + (decoded_visible_data[i + (width*height / 2)]);
     double recoverVal2 = decoded_visible_data[i] - (decoded_visible_data[i + (width*height / 2)]);
-    reconstructed_visible_data[i * 2] = recoverVal1;
-    reconstructed_visible_data[(i * 2) + 1] = recoverVal2;
+    reconstructed_visible_data[i * 2] = (int)recoverVal1;
+    reconstructed_visible_data[(i * 2) + 1] = (int)recoverVal2;
   }
   //	return dwtRecoveredImage;
 }
@@ -347,8 +327,7 @@ int* m3dvc::getDWTBlock(int* oneMergeFilter, int Width, int Height, int compress
   utility testTimer;
   //testTimer.startTimer();
 
-  for (int i = 0; i < (Width*Height / compressedIndicator); i++)
-  {
+  for (int i = 0; i < (Width*Height / compressedIndicator); i++) {
     int indexY = i / (Width / 2);
     int indexX = i % (Width / 2);
 
@@ -361,17 +340,14 @@ int* m3dvc::getDWTBlock(int* oneMergeFilter, int Width, int Height, int compress
 
 void m3dvc::M3DVC_depth_decode_2() {
   int expValIndex = 0;
-  for (int i = 0; i < (height*width) / 4; i++)
-  {
+  for (int i = 0; i < (height*width) / 4; i++) {
     int indexY = expValIndex / (width);
     int indexX = expValIndex % (width);
-    if (indexY % 2 != 0)
-    {
+    if (indexY % 2 != 0) {
       //			indexY++;
       expValIndex = width + expValIndex;
       i--;
-    } else
-    {
+    } else {
       decodedBlock[width*indexY + indexX] = encoded_depth_data[i];
       decodedBlock[width*indexY + indexX + 1] = encoded_depth_data[i];
       decodedBlock[width*(indexY + 1) + indexX] = encoded_depth_data[i];
@@ -387,16 +363,13 @@ void m3dvc::M3DVC_depth_encode_2(int* realDepthVal1D) {
 
   int returnValIndex = 0;
 
-  for (int i = 0; i < width*height; i = i + 2)
-  {
+  for (int i = 0; i < width*height; i = i + 2) {
     int indexY = i / (width);
     int indexX = i % (width);
-    if (indexY % 2 != 0)
-    {
+    if (indexY % 2 != 0) {
       //			indexY++;
       i = i + width - 2;
-    } else
-    {
+    } else {
       depthHolder = (int)(realDepthVal1D[i] + realDepthVal1D[i + 1] + realDepthVal1D[i + width] + realDepthVal1D[i + width + 1]) / 4;
       encoded_depth_data[returnValIndex] = depthHolder;
       returnValIndex++;
